requirements.txt
torch==2.2.0+cu121
-f https://download.pytorch.org/whl/cu121/torch_stable.html

pandas
numpy<2
transformers
Pillow
scikit-learn
xgboost
lightgbm
catboost
requests
opencv-python
textblob
matplotlib
seaborn
tqdm
wandb
aiohttp
aiofiles
asyncio
ipykernel